improvement(vercel): do not require setting the google-maps-api-key when it's not used
Previously, when you wanted to deploy the application to Vercel without setting preconfiguring it
with a Google Maps API key, then the user still had to supply the google-maps-api.key secret. Which
one cannot set to an empty value through the CLI. The only workaround was overwriting it through an
environment variable, which actually can be set to the empty string.

BREAKING CHANGE: Now one needs to copy now.template.json to now.json when you want to use the
google-maps-api-key secret for configuring the global Google Maps API key.